CFMM pools state resolution: WIP.
