+++ modules/dex-core/src/main/scala/org/ergoplatform/dex/domain/BoxInfo.scala
@@ -3,13 +3,13 @@ package org.ergoplatform.dex.domain
 import derevo.circe.{decoder, encoder}
 import derevo.derive
 import org.ergoplatform.ergo.BoxId
-import org.ergoplatform.ergo.models.ErgoBox
+import org.ergoplatform.ergo.models.Output
 import tofu.logging.derivation.loggable
 
 @derive(encoder, decoder, loggable)
-final case class BoxInfo(boxId: BoxId, value: Long)
+final case class BoxInfo(currentBoxId: BoxId, currentBoxValue: Long, lastConfirmedBoxGix: Long)
 
 object BoxInfo {
-  def fromBox(box: ErgoBox): BoxInfo =
-    BoxInfo(box.boxId, box.value)
+  def fromBox(box: Output): BoxInfo =
+    BoxInfo(box.boxId, box.value, box.globalIndex)
 }
