+++ modules/dex-core/src/main/scala/org/ergoplatform/ergo/models/Output.scala
@@ -12,6 +12,7 @@ final case class Output(
   transactionId: TxId,
   value: Long,
   index: Int,
+  globalIndex: Long,
   creationHeight: Int,
   settlementHeight: Int,
   ergoTree: SErgoTree,
